Augment Channel for other possible fields
The Channel struct is one of the "includes" fields for log entries.
Beyond having a type, it can also have several other fields that
provide information about the incident.

The Details field is particularly useful, as the first log entry for
every incident includes arbitrary data that can be sent by the service
opening the incident.