Allow package consumers to provide their own HTTP client
This updates the package to no longer make use of the `http.DefaultClient`, and
to instead use a package-local default client as well as to provide consumers
the ability to provide their own client in the time of need.

The reason for moving away from the `http.DefaultClient` is for two reasons:

1. It prevents other packages from being able to impact this one by making
   changes to the `http.DefaultClient`.

2. The isolation should help make this package more resilient, as we can tweak
   the client's transport settings to provide defaults appropriate for PagerDuty
   usage.

The package-local default client is not directly exposed, so as to hopefully
avoid people making out-of-band changes to its config and to instead provide
their own client.

This also supports the ability for consumers to provide their own HTTP client,
in the situation where the default client doesn't have the settings they need.
The ability to provide this client was done using an interface, `HTTPClient`, so
that consumers aren't forced to make use of `*http.Client` if they don't want
to.

Fixes #110

Signed-off-by: Tim Heckman <t@heckman.io>