Use pointer to Priority so don't send and empty priority for incident
Due to the changes to the Incident struct the library now inadvertently
sends an empty priority to the API (`"priority":{}`) and for
ManageIncidents this results in a response of:

    Priority id cannot be empty

To prevent sending an empty priority we can use a pointer to Priority.

Seems like a super simple fix for this - and as such I'm worried I've
missed something else.

Full discussion of this in this issue:

References: https://github.com/PagerDuty/go-pagerduty/issues/135