Update internal HTTP methods to not take pointer to map as argument
Map types themselves are reference types, and can be passed in as `nil` even
without declaring the value as a pointer to the map.

Likewise, for maps that are nil the `range` operator is safe and no-ops and so
we do not need a nil check before doing the range. If for some reason we wanted
a check, a length check would be more appropriate than a nil check but it's not
necessary.