Fix possible nil pointer dereference in APIError.NotFound() method
If someone were to receive a PagerDuty response that did not contain the error
JSON object, they would hit a nil pointer dereference if they attempted to call
the .NotFound() method if the status code is not 404.

This fixes it by doing a nil check before attempting to read the value to do the
comparison.