Fix overlapping struct fields & last golint errors
There have been a few issues filed over the past few years, relative to
unexpected or unreliable behavior in the client, due to overlapping struct
fields because of an embedded type or conflicting JSON field tags. This could
make it appear like some fields are unset, because `encoding/json` has put the
value in the overlapped field and not the one the programmer is accessing.
Unfortunately, fixing these issues are technically a breaking API change and so
we waited to completely fix them until now.

In the spirit of making the struct fields consistently named, this also updates
the `HttpStatus` field in the `EventResponse` struct to be `HTTPStatus`. This is
purely cosmetic, but feels appropriate to do now since the removal if the `Id`
field makes it the last one.

Fixes #218
Fixes #316
Fixes #268