Update APIError.Error() to provide more helpful error messages
While refactoring the API handling with the `APIError` type, we gave callers
more capability to inspect the errors that come from the API so that we can
handle them. Unfortunately, in the process we lost some fidelity in the strings
returned by its `Error()` method, compared to how errors used to be handled.

This updates the `APIError.Error()` method to now include the first detailed
error string provided from the API, if there is one, and incidate how many more
existed in the response (without printing them). This should help keep the error
tidy, while making it actionable by consumers. Consumers of this package are
still encouraged to deeply inspect the error, to handle each one.

This was inspired by #327, which the reporter closed after learning how to
deeply inspect the returned error value.