Add support for adding email filters for Generic Email Integrations
This change adds support for adding different email filter rules to a generic
inbound email integration associated with a specific Service. This is done by
adding two new fields to the Integration struct.

While implementing this functionality I found that the API documentation from
PagerDuty lacked details about what the correct values are for specific fields,
and specifically what each value means. In fact, I actually had to go into the
PagerDuty UI to try and set up email filters, so that I could guess what the
different API values actually meant.

Because of that, I chose to create an enum-like type for each thing that wasn't
well documented to make it easier to consume via this client. That's resulted in
quite a lot of extra code, but the result is that anyone who wishes to consume
this area of the Go SDK should be able to figure it out how to use it without
much trouble.

Because of how the code grew, I also chose to move the Service
Integration-specific code into a new file (service_integration.go).

Closes #315