Backport PR #382 to release-1.4.x branch
The bug this fixes was introduced as part of v1.4.0, so we need to backport it
into the v1.4.x release line. Because the master branch has diverged towards
v1.5.0, we need to manually backport this to the release-1.4.x branch.

Original commit message (0933613394cf37960b82c3fc0c1b7285caae122e):

The PagerDuty REST API documentation[1] indicates that the errors field of an
error response is an array of error strings. However, as shown in #339 the API
sometimes violates that specification and instead returns the errors field as
a string.

There is a PagerDuty customer support ticket open[2] to to address this issue, but
there is currently no ETA on the resolution. As such we are going to create this
workaround to properly parse the invalid responses and return the proper error
type to callers.

Closes #339
Backports #382

[1] https://developer.pagerduty.com/docs/ZG9jOjExMDI5NTYz-errors
[2] https://tickets.pagerduty.com/hc/requests/341221