Support adding conference bridge when creating or managing incidents
Up until now you could view the conference bridge associated with an incident
when viewing its details via this package, but you could not add a conference
bridge to a new or existing incident. This change adds the ability to include a
conference bridge when creating a new incident, or when you're updating an
existing one.

Closes #373