Mark Type struct field deprecated, for incident creation + management
We should not expect consumers to set this value, as there is only a single
value that is valid (`incident`) based on how we use the API. As such, let's
deprecate the field and just set it in the methods that need it.

This also updates the tests to assert that we're automatically setting it
appropriately.

Closes #390